requirements.txt
flask
flask-cors
flask-mail
uuid
python-dotenv
email-validator